add systemt events and actions to example.future.config
