+++ packages/admin/example.future.config.ts
@@ -1,7 +1,16 @@
-import { Config, createFramework } from 'core';
+import { Config, createFramework, extractSchemaOptions } from 'core';
 import { GoogleIntegration } from 'future-google';
 import { MailchimpIntegration } from 'future-mailchimp';
 import { SlackIntegration } from 'future-slack';
+
+import {
+  BASE_RECORD_SCHEMA,
+  CREATE_NOTE_OUTPUT_SCHEMA,
+  CREATE_NOTE_SCHEMA,
+  CREATE_TASK_OUTPUT_SCHEMA,
+  CREATE_TASK_SCHEMA,
+  RECORD_SCHEMA,
+} from '@/domains/workflows/constants';
 
 // // We have an admin db
 // // Enter secrets and shit it saves it to admin db for that integration
@@ -31,8 +40,107 @@ export const REDIRECT_URI = new URL(redirectPath, redirectHost).toString();
 export const config: Config = {
   name: 'kepler',
   //logConfig: {}, // TODO: Add this
-  systemActions: [],
-  systemEvents: [],
+  systemActions: [
+    {
+      pluginName: 'system',
+      type: 'CREATE_NOTE',
+      label: 'Create Note',
+      icon: {
+        alt: 'Create Note',
+        icon: 'plus-icon',
+      },
+      category: 'NOTE',
+      description: 'Create a new note',
+      schema: CREATE_NOTE_SCHEMA as any,
+      async getSchemaOptions() {
+        const options = extractSchemaOptions({ schema: CREATE_NOTE_SCHEMA });
+        return options;
+      },
+      outputSchema: CREATE_NOTE_OUTPUT_SCHEMA,
+      executor: async () => {
+        //executor
+      },
+    },
+    {
+      pluginName: 'system',
+      type: 'CREATE_TASK',
+      label: 'Create Task',
+      icon: {
+        alt: 'Create Task',
+        icon: 'plus-icon',
+      },
+      category: 'NOTE',
+      description: 'Create a new task',
+      schema: CREATE_TASK_SCHEMA as any,
+      async getSchemaOptions() {
+        const options = extractSchemaOptions({ schema: CREATE_TASK_SCHEMA });
+        return options;
+      },
+      outputSchema: CREATE_TASK_OUTPUT_SCHEMA,
+      executor: async () => {
+        //executor
+      },
+    },
+  ],
+  systemEvents: [
+    {
+      key: 'record_created',
+      schema: BASE_RECORD_SCHEMA,
+      triggerProperties: {
+        type: 'RECORD_CREATED',
+        label: 'Record Created',
+        icon: {
+          alt: 'Record Created',
+          icon: 'record-created',
+        },
+        description: 'Triggered when a record is created',
+        schema: BASE_RECORD_SCHEMA,
+        async getSchemaOptions() {
+          const options = extractSchemaOptions({ schema: BASE_RECORD_SCHEMA });
+          return options;
+        },
+        outputSchema: RECORD_SCHEMA,
+      },
+    },
+    {
+      key: 'record_updated',
+      schema: BASE_RECORD_SCHEMA,
+      triggerProperties: {
+        type: 'RECORD_UPDATED',
+        label: 'Record Updated',
+        icon: {
+          alt: 'Record Updated',
+          icon: 'edit',
+        },
+        description: 'Triggered when a record is updated',
+        schema: BASE_RECORD_SCHEMA,
+        async getSchemaOptions() {
+          const options = extractSchemaOptions({ schema: BASE_RECORD_SCHEMA });
+          return options;
+        },
+        outputSchema: RECORD_SCHEMA,
+      },
+    },
+    {
+      key: 'record_deleted',
+      schema: BASE_RECORD_SCHEMA,
+      triggerProperties: {
+        type: 'RECORD_DELETED',
+        label: 'Record Deleted',
+        icon: {
+          alt: 'Record Deleted',
+          icon: 'trash',
+        },
+        description: 'Triggered when a record is deleted',
+        schema: BASE_RECORD_SCHEMA,
+        async getSchemaOptions() {
+          const options = extractSchemaOptions({ schema: BASE_RECORD_SCHEMA });
+          return options;
+        },
+        outputSchema: RECORD_SCHEMA,
+      },
+    },
+  ],
   plugins: [
     new MailchimpIntegration({
       config: {
